Add files via upload
Seistronix Ras24 Seismic evaluation The software is not fully compatible with Windows 10 x64 and is incomplete install. We have made the installation complete rather than compatibility settings.  If you have encountered this problem you can find the tool below. geophysical studies such as masw seismic interpretation of the evaluation software for the RAS24 v2.15 64-bit architecture after the installation, we encounter errors such as file failure. Initially I needed a 32 bit system and so it was an old windows in the virtual machine eg: working on XP was a waste of time and more laborious. Again, from such a scenario to the actual physical up-to-date system, we encounter errors and problems when needed or attempted. This time we can overcome this problem as a result of our efforts to achieve a result that we share with you. And so we offer a facilitator tool.